+++ src/CostPlan/TableSingle.java
@@ -16,7 +16,7 @@ public class TableSingle extends TableSet2 {
 
 	private OneTableOperator op;
 	
-	public TableSingle(String table, OneTableOperator op, HashMap<vWrapper, Integer> vVals, UnionFind union) {
+	public TableSingle(String table, OneTableOperator op, HashMap<vWrapper, Long> vVals, UnionFind union) {
 		super(vVals, union);
 		
 		this.cost = 0;
@@ -29,12 +29,12 @@ public class TableSingle extends TableSet2 {
 	}
 	
 	@Override
-	public int vValCompute(String attr) {
+	public long vValCompute(String attr) {
 		vWrapper key = new vWrapper(this,attr);
 		if (vVals.containsKey(key))
 			return vVals.get(key);
 		
-		int v;
+		long v;
 		if (op instanceof ScanOperator)
 			v = vValBase(attr, (ScanOperator) op);
 		else if (op instanceof IndexScanOperator)
@@ -45,7 +45,7 @@ public class TableSingle extends TableSet2 {
 		return v;
 	}
 	
-	private int computeSize() {
+	private long computeSize() {
 		if (op instanceof ScanOperator)
 			return sizeBase((ScanOperator) op);
 		else if (op instanceof IndexScanOperator)
@@ -54,11 +54,11 @@ public class TableSingle extends TableSet2 {
 			return sizeSelect((SelectOperator) op);
 	}
 	
-	private int sizeBase(ScanOperator scan) {
+	private long sizeBase(ScanOperator scan) {
 		return Math.max(scan.getTableInfo().getNumTuples(), 1);
 	}
 	
-	private int sizeSelect(SelectOperator slct) {
+	private long sizeSelect(SelectOperator slct) {
 		TableInfo tableInfo = slct.getTableInfo();
 		
 		HashMap<String, Pair> selectRange = slct.getSelectRange();
@@ -67,50 +67,50 @@ public class TableSingle extends TableSet2 {
 		for (Entry<String, Pair> entry : selectRange.entrySet()) {
 			String attr = entry.getKey();
 			String col = attr.substring(attr.indexOf('.') + 1);
-			int max = tableInfo.getColumns().get(col).max;
-			int min = tableInfo.getColumns().get(col).min;
+			long max = tableInfo.getColumns().get(col).max;
+			long min = tableInfo.getColumns().get(col).min;
 			
-			int range = Math.min(max, entry.getValue().high) - Math.max(min, entry.getValue().low) + 1;
+			long range = Math.min(max, entry.getValue().high) - Math.max(min, entry.getValue().low) + 1;
 			
 			reduction *= ((double) range) / (max - min + 1);
 		}
 		
-		return Math.max(1, (int) (tableInfo.getNumTuples() * reduction));
+		return Math.max(1, (long) (tableInfo.getNumTuples() * reduction));
 	}
 	
-	private int sizeSelect(IndexScanOperator idx) {
+	private long sizeSelect(IndexScanOperator idx) {
 		String attr = idx.indexAttribute();
 		String col = attr.substring(attr.indexOf('.') + 1);
 		
 		TableInfo tableInfo = idx.getTableInfo();
-		int max = tableInfo.getColumns().get(col).max;
-		int min = tableInfo.getColumns().get(col).min;
+		long max = tableInfo.getColumns().get(col).max;
+		long min = tableInfo.getColumns().get(col).min;
 		
-		int range = Math.min(max, idx.getHighkey()) - Math.max(min, idx.getLowkey()) + 1;
+		long range = Math.min(max, idx.getHighkey()) - Math.max(min, idx.getLowkey()) + 1;
 		double reduction = ((double) range) / (max - min + 1);
 		
-		return Math.max(1, (int) (tableInfo.getNumTuples() * reduction));
+		return Math.max(1, (long) (tableInfo.getNumTuples() * reduction));
 	}
 	
-	private int vValBase(String attr, ScanOperator scan) {
+	private long vValBase(String attr, ScanOperator scan) {
 		TableInfo tableInfo = scan.getTableInfo();
 		String col = attr.substring(attr.indexOf('.') + 1);
 		
-		int max = tableInfo.getColumns().get(col).max;
-		int min = tableInfo.getColumns().get(col).min;
+		long max = tableInfo.getColumns().get(col).max;
+		long min = tableInfo.getColumns().get(col).min;
 		
-		int result = max - min + 1;
+		long result = max - min + 1;
 		
 		return Math.max(Math.min(result, tableInfo.getNumTuples()), 1);
 	}
 	
-	private int vValSelect(String attr, SelectOperator slct) {
+	private long vValSelect(String attr, SelectOperator slct) {
 		TableInfo tableInfo = slct.getTableInfo();
 		String col = attr.substring(attr.indexOf('.') + 1);
-		int max = tableInfo.getColumns().get(col).max;
-		int min = tableInfo.getColumns().get(col).min;
+		long max = tableInfo.getColumns().get(col).max;
+		long min = tableInfo.getColumns().get(col).min;
 		
-		int result;
+		long result;
 		if (slct.getSelectRange().get(attr) != null)
 			result = Math.min(max, slct.getSelectRange().get(attr).high) - Math.max(min, slct.getSelectRange().get(attr).low) + 1;
 		else
@@ -119,13 +119,13 @@ public class TableSingle extends TableSet2 {
 		return Math.max(Math.min(result, tableInfo.getNumTuples()), 1);
 	}
 	
-	private int vValSelect(String attr, IndexScanOperator idx) {
+	private long vValSelect(String attr, IndexScanOperator idx) {
 		TableInfo tableInfo = idx.getTableInfo();
 		String col = attr.substring(attr.indexOf('.') + 1);
-		int max = tableInfo.getColumns().get(col).max;
-		int min = tableInfo.getColumns().get(col).min;
+		long max = tableInfo.getColumns().get(col).max;
+		long min = tableInfo.getColumns().get(col).min;
 		
-		int result;
+		long result;
 		if (idx.indexAttribute().equals(attr))
 			result = Math.min(max, idx.getHighkey()) - Math.max(min, idx.getLowkey()) + 1;
 		else
